requirements.txt
# Core ML & Data Processing
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
joblib>=1.0.0

# Image Processing & OCR
opencv-python>=4.5.0
Pillow>=8.0.0
pytesseract>=0.3.8
easyocr>=1.6.0

# FastAPI Backend
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-multipart>=0.0.5
aiofiles>=22.0.0

# Streamlit Frontend
streamlit>=1.28.0

# GenAI Integration
groq>=0.3.0
openai>=1.0.0

# Enhanced ML
xgboost>=1.6.0
lightgbm>=3.3.0
imbalanced-learn>=0.10.0

# Utilities
python-dotenv>=0.19.0
pydantic>=2.0.0
requests>=2.25.0

# Optional Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0
